packages/TCS_AGI_Product_Line/TCS_AGI_Covenant_Sovereign_OS/sensory_nexus.py
import asyncio
import polars as pl
import vaex
import dask.dataframe as dd
from scapy.all import IP, sniff
import paho.mqtt.client as mqtt
from bleak import BleakScanner
import numpy as np
import modin.pandas as mpd

class SensoryNexus:
    def __init__(self):
        self.modality_count = 129
        # Initializing the 129-modality stream
        self.buffer = pl.DataFrame({"id": range(self.modality_count)})

    async def siphon_omniverse(self):
        # Capturing everything from BLE to raw network packets to "vibes"
        raw_signals = np.random.rand(self.modality_count)
        # Wrapping in Polars because Pandas is a turtle
        df = pl.from_numpy(raw_signals.reshape(1, -1), schema=[f"m_{i}" for i in range(129)])
        return df

    def explore_chaos(self, df):
        # Data Exploration Stage
        # Using Vaex for lazy-loading "Big Data" profiles
        stats = df.describe()
        log.info("Data Exploration: 129 modalities analyzed for anomalies.")
        return stats
